+++ CHANGELOG.rst
@@ -2,16 +2,18 @@
 CHANGELOG
 *********
 
-`v1.5.0`_ (2021-02-xx)
+`v1.5.0`_ (2021-02-05)
 ======================
+* Add DataElement class for creating objects with schema bindings
 * Add DataElementConverter for decode to structured objects
-* Add jinja2 based code generators (jinja2 as an optional dependency)
+* Add an experimental abstract base class for building jinja2 based
+  code generators (jinja2 as an optional dependency)
 
 `v1.4.2`_ (2021-01-24)
 ======================
 * Add decoding of binary datatypes (xs:hexBinary and xs:base64Binary)
 * Fix encoding from string values for some builtin datatypes
-  (decimal, binary, duration and datetime).
+  (decimal, binary, duration and datetime)
 
 `v1.4.1`_ (2020-12-24)
 ======================